Update changes to linter files
