+++ js/info.js
@@ -1,72 +1,72 @@
 const data = [
-    {
-      id: 1,
-      image: './images/eerste.svg',
-      name: 'Tonic',
-      description: {
-        type: 'CANOPY',
-        interface: 'Backend Dev',
-        year: 2015,
-      },
-      fullDesciption:
+  {
+    id: 1,
+    image: './images/eerste.svg',
+    name: 'Tonic',
+    description: {
+      type: 'CANOPY',
+      interface: 'Backend Dev',
+      year: 2015,
+    },
+    fullDesciption:
         'Lorem ipsum dolor sit amet consectetur adipisicing elit. Repudiandae eius quae optio et quas, quibusdam quis excepturi omnis? Quos, officia accusamus. Aperiam, molestiae elit. atque maiores aperiam.    Lorem, ipsum dolor sit amet consectetur adipisicing elit. Expedita ex quam officia ea dolor. Ea quasi in id dolorum ',
-      detail:
+    detail:
         'A daily selection of privately personalized reads; no accounts or sign-ups required.',
-      technologies: ['Html', 'Css', 'Javascript', 'Ruby on Rails'],
-      liveDemo: 'https://github.com/Athaliah8',
-      sourceCode: 'https://github.com/Athaliah8',
+    technologies: ['Html', 'Css', 'Javascript', 'Ruby on Rails'],
+    liveDemo: 'https://github.com/Athaliah8',
+    sourceCode: 'https://github.com/Athaliah8',
+  },
+  {
+    id: 2,
+    image: './images/eerste.svg',
+    name: 'Multi-Post Stories',
+    description: {
+      type: 'CANOPY',
+      interface: 'Backend Dev',
+      year: 2015,
     },
-    {
-      id: 2,
-      image: './images/eerste.svg',
-      name: 'Multi-Post Stories',
-      description: {
-        type: 'CANOPY',
-        interface: 'Backend Dev',
-        year: 2015,
-      },
-      fullDesciption:
+    fullDesciption:
         'Lorem Ipsum is simply dummy text of the printing and typesetting industry. Lorem Ipsum has been the industrys standard dummy text ever since the 1500s, when an unknown printer took a galley of type and scrambled it to make a type specimen book. It has survived not only five centuries, but also the leap into electronic typesetting, remaining essent',
-      detail:
+    detail:
         'A smart assistant to make driving more safe, efficient, and fun by unlocking your most expensive computer: your car.',
-      technologies: ['Html', 'Css', 'Javascript', 'Ruby on Rails'],
-      liveDemo: 'https://athaliah8.github.io/Mobile_Portfolio/',
-      sourceCode: 'https://github.com/Athaliah8',
+    technologies: ['Html', 'Css', 'Javascript', 'Ruby on Rails'],
+    liveDemo: 'https://athaliah8.github.io/Mobile_Portfolio/',
+    sourceCode: 'https://github.com/Athaliah8',
+  },
+  {
+    id: 3,
+    image: './images/eerste.svg',
+    name: 'Tonic',
+    description: {
+      type: 'CANOPY',
+      interface: 'Backend Dev',
+      year: 2015,
     },
-    {
-      id: 3,
-      image: './images/eerste.svg',
-      name: 'Tonic',
-      description: {
-        type: 'CANOPY',
-        interface: 'Backend Dev',
-        year: 2015,
-      },
-      fullDesciption:
+    fullDesciption:
         'Lorem Ipsum is simply dummy text of the printing and typesetting industry. Lorem Ipsum has been the industrys standard dummy text ever since the 1500s, when an unknown printer took a galley of type and scrambled it to make a type specimen book. It has survived not only five centuries, but also the leap into electronic typesetting, remaining essent',
-      detail:
+    detail:
         'Exploring the future of media in Facebooks first Virtual Reality app; a place to discover and enjoy 360 photos and videos on Gear VR.',
-      technologies: ['Html', 'Css', 'Javascript', 'Ruby on Rails' ],
-      liveDemo: 'https://athaliah8.github.io/Mobile_Portfolio/',
-      sourceCode: 'https://github.com/Athaliah8',
+    technologies: ['Html', 'Css', 'Javascript', 'Ruby on Rails'],
+    liveDemo: 'https://athaliah8.github.io/Mobile_Portfolio/',
+    sourceCode: 'https://github.com/Athaliah8',
+  },
+  {
+    id: 4,
+    image: './images/eerste.svg',
+    name: 'Multi-Post Stories',
+    description: {
+      type: 'CANOPY',
+      interface: 'Backend Dev',
+      year: 2015,
     },
-    {
-      id: 4,
-      image: './images/eerste.svg',
-      name: 'Multi-Post Stories',
-      description: {
-        type: 'CANOPY',
-        interface: 'Backend Dev',
-        year: 2015,
-      },
-      fullDesciption:
+    fullDesciption:
         'Lorem Ipsum is simply dummy text of the printing and typesetting industry. Lorem Ipsum has been the industrys standard dummy text ever since the 1500s, when an unknown printer took a galley of type and scrambled it to make a type specimen book. It has survived not only five centuries, but also the leap into electronic typesetting, remaining essent',
-      detail:
+    detail:
         'Exploring the future of media in Facebooks first Virtual Reality app; a place to discover and enjoy 360 photos and videos on Gear VR.',
-      technologies: ['Html', 'Css', 'Javascript', 'Ruby on Rails'],
-      liveDemo: 'https://athaliah8.github.io/Mobile_Portfolio/',
-      sourceCode: 'https://github.com/Athaliah8',
-    },
-   
-  ];
-  export default data;
+    technologies: ['Html', 'Css', 'Javascript', 'Ruby on Rails'],
+    liveDemo: 'https://athaliah8.github.io/Mobile_Portfolio/',
+    sourceCode: 'https://github.com/Athaliah8',
+  },
+
+];
+export default data;
