+++ script.js
@@ -33,33 +33,31 @@ function outputMessage(accepted) {
   }
 }
 
-
-
 // Get the modal
-var modal = document.getElementById("myModal");
+const modal = document.getElementById('myModal');
 
 // Get the button that opens the modal
-var btn = document.getElementsByClassName(".see-btn");
+const btn = document.getElementsByClassName('.see-btn');
 
 // Get the <span> element that closes the modal
-var span = document.getElementsByClassName("close")[0];
+const span = document.getElementsByClassName('close')[0];
 
-// When the user clicks the button, open the modal 
-btn.onclick = function() {
-  modal.style.display = "block";
-}
+// When the user clicks the button, open the modal
+btn.onclick = function () {
+  modal.style.display = 'block';
+};
 
 // When the user clicks on <span> (x), close the modal
-span.onclick = function() {
-  modal.style.display = "none";
-}
+span.onclick = function () {
+  modal.style.display = 'none';
+};
 
 // When the user clicks anywhere outside of the modal, close it
-window.onclick = function(event) {
+window.onclick = function (event) {
   if (event.target == modal) {
-    modal.style.display = "none";
+    modal.style.display = 'none';
   }
-}
+};
 
 const form = document.querySelector('.container-form');
 
@@ -79,5 +77,3 @@ form.addEventListener('submit', (element) => {
   }
   outputMessage(true);
 });
-
-
